Backend/requirements.txt
fastapi
pydantic
pandas
transformers
uvicorn
torch
google-generativeai